update debug removeAssets (cache/0/debug/1)

+++ actions/RemoveAssetsAction.php
@@ -4,23 +4,40 @@
  */
 class RemoveAssetsAction extends CAction
 {
+	private $_debug;
+	
 	private function rmDirRec($dir)
 	{
 		$objs = glob($dir."/*");
 		if ($objs){
 			foreach($objs as $obj) {
-				is_dir($obj) ? $this->rmDirRec($obj) : unlink($obj);
+				if (is_dir($obj)) {
+					$this->rmDirRec($obj);
+				} else {	
+					if ($this->_debug) {
+						echo 'file: '.$obj.'<br />';
+					} else {
+						unlink($obj);
+					}
+				}
 			}
 		}
-		rmdir($dir);
+		if ($this->_debug) {
+			echo 'directory: '.$obj.'<br />';			
+		} else {
+			rmdir($dir);
+		}	
 	}
 
 	
-	public function run($clearCache = false)
+	public function run($clearCache = 0, $debug = 0)
 	{
 		defined('DS') or define('DS',DIRECTORY_SEPARATOR);		
-		
+		$this->_debug = $debug;
 		$AM  = new CAssetManager;
+		if ($this->_debug) {
+			echo 'basePath: '.$AM->getBasePath().'<br />';
+		}
 		$dir = $AM->getBasePath();
 		if(file_exists($dir)) {
 			$files = glob($dir.DS."*");
@@ -29,7 +46,11 @@ class RemoveAssetsAction extends CAction
 				if(is_dir($del) && ($clearCache || !($s == 'cache'))) {
 					$this->rmDirRec($del);
 				} elseif (($s != 'cache')) {
-					unlink($del);
+					if ($this->_debug) {
+						echo 'file: '.$del.'<br />';
+					} else {
+						unlink($del);
+					}	
 				}
 			}
 			echo 'all files have been removed';
